% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata_soccer.R
\name{soccer.data.txt}
\alias{soccer.data.txt}
\title{Dataset soccer.data.txt}
\format{This data set contains 2 columns. The first column is the 
football season year. The second column is the average goals-per-game 
in each season. Data are derived from all English professional league 
soccer results from 1888-2014 (engsoccerdata R package).}
\source{
James P. Curley, engsoccerdata: English Soccer Data 1871-
2016. R package version 0.1.5 (2016), doi: 10.5281/zenodo.13158.
}
\description{
This data set is a small time series to test the \code{ACA} package.
}
